Pessoal/Estudos/Plataforma de Cursos/requirements.txt
chainlit
fastapi
openai
pydantic
python-dotenv
requests
streamlit
uvicorn